Fixes for "warning: subscript has type `char'" The problem is that isupper and tolower are macros under solaris.

+++ trunk/ptclassic/src/octtools/Packages/tap/tap.c
@@ -319,7 +319,7 @@ char *mode;
     /* create technology view name */
     viewStr = util_strsav(view);
     for (cPtr = viewStr; *cPtr; cPtr++) {
-	if (isupper(*cPtr)) *cPtr = tolower(*cPtr);
+	if (isupper((int)*cPtr)) *cPtr = tolower((int)*cPtr);
     }
 
     /* open technology facet */
